Slight tweak to the type of the `cond` Expr variant
Changed from `Vec<Node<(Expr, Expr)>>` to Vec<(Node<Expr>, Node<Expr>)>

This makes it more consistent with the other enum variants.

Also, while it's possible to go from the old form to the new form, it's
not possible to go the other way, since you can't move out of an Rc.